requirements.txt
beautifulsoup4==4.6.0
PySocks==1.6.8
requests==2.18.4
urllib3==1.22

